// src/domains/service-clients/course/proto/generated/course/types/enrollment.ts
export interface EnrollmentData {
  id: string;
  userId: string;
  courseId: string;
  status: string;
  progress: number;
  enrolledAt: string;
  completedAt: string;
  createdAt: string;
  updatedAt: string;
  deletedAt?: string | undefined;
  course: EnrollmentCourseData | undefined;
}

export interface EnrollmentCourseData {
  id: string;
  title: string;
  rating: number;
  thumbnail: string;
  category: string;
  level: string;
  lessonsCount: number;
  instructor: User | undefined;
}

/** Enrollment Messages */
export interface EnrollmentDetail {
  enrollmentId: string;
  userId: string;
  courseId: string;
  progressPercent: number;
  status: string;
  enrolledAt: string;
  sections: EnrollmentSection[];
}

export interface EnrollmentSection {
  id: string;
  title: string;
  description?: string | undefined;
  order: number;
  isPublished: boolean;
  lessons: LessonProgress[];
  quiz?: QuizProgress | undefined;
}

export interface LessonProgress {
  id: string;
  title: string;
  order: number;
  /** second ?s = 1 */
  duration?: number | undefined;
  completed: boolean;
  completedAt?: string | undefined;
}

export interface QuestionOption {
  value: string;
}

export interface QuizQuestion {
  id: string;
  requirePassingScore: boolean;
  options: QuestionOption[];
  timeLimit?: number | undefined;
  question: string;
  type: string;
  explanation?: string | undefined;
  score?: number | undefined;
  correctAnswer?: string | undefined;
}
// ...
export const EnrollmentCourseData: MessageFns<EnrollmentCourseData> = {
  encode(message: EnrollmentCourseData, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.id !== "") {
      writer.uint32(10).string(message.id);
    }
    if (message.title !== "") {
      writer.uint32(18).string(message.title);
    }
    if (message.rating !== 0) {
      writer.uint32(24).int32(message.rating);
    }
    if (message.thumbnail !== "") {
      writer.uint32(34).string(message.thumbnail);
    }
    if (message.category !== "") {
      writer.uint32(42).string(message.category);
    }
    if (message.level !== "") {
      writer.uint32(50).string(message.level);
    }
    if (message.lessonsCount !== 0) {
      writer.uint32(56).int32(message.lessonsCount);
    }
    if (message.instructor !== undefined) {
      User.encode(message.instructor, writer.uint32(66).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): EnrollmentCourseData {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseEnrollmentCourseData();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.id = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.title = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.rating = reader.int32();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.thumbnail = reader.string();
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.category = reader.string();
          continue;
        }
        case 6: {
          if (tag !== 50) {
            break;
          }

          message.level = reader.string();
          continue;
        }
        case 7: {
          if (tag !== 56) {
            break;
          }

          message.lessonsCount = reader.int32();
          continue;
        }
        case 8: {
          if (tag !== 66) {
            break;
          }

          message.instructor = User.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): EnrollmentCourseData {
    return {
      id: isSet(object.id) ? globalThis.String(object.id) : "",
      title: isSet(object.title) ? globalThis.String(object.title) : "",
      rating: isSet(object.rating) ? globalThis.Number(object.rating) : 0,
      thumbnail: isSet(object.thumbnail) ? globalThis.String(object.thumbnail) : "",
      category: isSet(object.category) ? globalThis.String(object.category) : "",
      level: isSet(object.level) ? globalThis.String(object.level) : "",
      lessonsCount: isSet(object.lessonsCount) ? globalThis.Number(object.lessonsCount) : 0,
      instructor: isSet(object.instructor) ? User.fromJSON(object.instructor) : undefined,
    };
  },

  toJSON(message: EnrollmentCourseData): unknown {
    const obj: any = {};
    if (message.id !== "") {
      obj.id = message.id;
    }
    if (message.title !== "") {
      obj.title = message.title;
    }
    if (message.rating !== 0) {
      obj.rating = Math.round(message.rating);
    }
    if (message.thumbnail !== "") {
      obj.thumbnail = message.thumbnail;
    }
    if (message.category !== "") {
      obj.category = message.category;
    }
    if (message.level !== "") {
      obj.level = message.level;
    }
    if (message.lessonsCount !== 0) {
      obj.lessonsCount = Math.round(message.lessonsCount);
    }
    if (message.instructor !== undefined) {
      obj.instructor = User.toJSON(message.instructor);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<EnrollmentCourseData>, I>>(base?: I): EnrollmentCourseData {
    return EnrollmentCourseData.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<EnrollmentCourseData>, I>>(object: I): EnrollmentCourseData {
    const message = createBaseEnrollmentCourseData();
    message.id = object.id ?? "";
    message.title = object.title ?? "";
    message.rating = object.rating ?? 0;
    message.thumbnail = object.thumbnail ?? "";
    message.category = object.category ?? "";
    message.level = object.level ?? "";
    message.lessonsCount = object.lessonsCount ?? 0;
    message.instructor = (object.instructor !== undefined && object.instructor !== null)
      ? User.fromPartial(object.instructor)
      : undefined;
    return message;
  },
};
// ...
export const EnrollmentSection: MessageFns<EnrollmentSection> = {
  encode(message: EnrollmentSection, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.id !== "") {
      writer.uint32(10).string(message.id);
    }
    if (message.title !== "") {
      writer.uint32(18).string(message.title);
    }
    if (message.description !== undefined) {
      writer.uint32(26).string(message.description);
    }
    if (message.order !== 0) {
      writer.uint32(32).int32(message.order);
    }
    if (message.isPublished !== false) {
      writer.uint32(40).bool(message.isPublished);
    }
    for (const v of message.lessons) {
      LessonProgress.encode(v!, writer.uint32(58).fork()).join();
    }
    if (message.quiz !== undefined) {
      QuizProgress.encode(message.quiz, writer.uint32(74).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): EnrollmentSection {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseEnrollmentSection();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.id = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.title = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.description = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 32) {
            break;
          }

          message.order = reader.int32();
          continue;
        }
        case 5: {
          if (tag !== 40) {
            break;
          }

          message.isPublished = reader.bool();
          continue;
        }
        case 7: {
          if (tag !== 58) {
            break;
          }

          message.lessons.push(LessonProgress.decode(reader, reader.uint32()));
          continue;
        }
        case 9: {
          if (tag !== 74) {
            break;
          }

          message.quiz = QuizProgress.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): EnrollmentSection {
    return {
      id: isSet(object.id) ? globalThis.String(object.id) : "",
      title: isSet(object.title) ? globalThis.String(object.title) : "",
      description: isSet(object.description) ? globalThis.String(object.description) : undefined,
      order: isSet(object.order) ? globalThis.Number(object.order) : 0,
      isPublished: isSet(object.isPublished)
        ? globalThis.Boolean(object.isPublished)
        : isSet(object.is_published)
        ? globalThis.Boolean(object.is_published)
        : false,
      lessons: globalThis.Array.isArray(object?.lessons)
        ? object.lessons.map((e: any) => LessonProgress.fromJSON(e))
        : [],
      quiz: isSet(object.quiz) ? QuizProgress.fromJSON(object.quiz) : undefined,
    };
  },

  toJSON(message: EnrollmentSection): unknown {
    const obj: any = {};
    if (message.id !== "") {
      obj.id = message.id;
    }
    if (message.title !== "") {
      obj.title = message.title;
    }
    if (message.description !== undefined) {
      obj.description = message.description;
    }
    if (message.order !== 0) {
      obj.order = Math.round(message.order);
    }
    if (message.isPublished !== false) {
      obj.isPublished = message.isPublished;
    }
    if (message.lessons?.length) {
      obj.lessons = message.lessons.map((e) => LessonProgress.toJSON(e));
    }
    if (message.quiz !== undefined) {
      obj.quiz = QuizProgress.toJSON(message.quiz);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<EnrollmentSection>, I>>(base?: I): EnrollmentSection {
    return EnrollmentSection.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<EnrollmentSection>, I>>(object: I): EnrollmentSection {
    const message = createBaseEnrollmentSection();
    message.id = object.id ?? "";
    message.title = object.title ?? "";
    message.description = object.description ?? undefined;
    message.order = object.order ?? 0;
    message.isPublished = object.isPublished ?? false;
    message.lessons = object.lessons?.map((e) => LessonProgress.fromPartial(e)) || [];
    message.quiz = (object.quiz !== undefined && object.quiz !== null)
      ? QuizProgress.fromPartial(object.quiz)
      : undefined;
    return message;
  },
};
// ...
export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}