// src/domains/service-clients/course/proto/generated/course/types/review.ts
export interface ReviewData {
  id: string;
  userId: string;
  courseId: string;
  enrollmentId: string;
  rating: number;
  comment: string;
  createdAt: string;
  updatedAt: string;
  /** optional string deleted_at = 8; */
  user: User | undefined;
}